hw4

+++ models/contacts.js
@@ -1,23 +1,30 @@
 const Contact = require('./mongoosSchema');
+// const userId = req.user.id;
 
-const listContacts = async () => {
-  return await Contact.find();
+const listContacts = async (req) => {
+  const userId = req.user.id;
+  console.log(userId);
+  return await Contact.find({ ownerId: userId });
 };
 
-const getContactById = async (contactId) => {
-  return await Contact.findById(contactId);
+const getContactById = async (req, contactId) => {
+  const userId = req.user.id;
+  return await Contact.findById(contactId).where({ ownerId: userId });
 };
 
-const removeContact = async (contactId) => {
-  return await Contact.findByIdAndRemove(contactId);
+const removeContact = async (req, contactId) => {
+  const userId = req.user.id;
+  return await Contact.findByIdAndRemove(contactId).where({ ownerId: userId });
 };
 
-const addContact = async (body) => {
-  return await Contact.create(body);
+const addContact = async (req, body) => {
+  const userId = req.user.id;
+  return await Contact.create({...body, ownerId: userId});
 };
 
-const updateContact = async (contactId, body) => {
-  return await Contact.findByIdAndUpdate(contactId, body, { new: true });
+const updateContact = async (req, contactId, body) => {
+  const userId = req.user.id;
+  return await Contact.findByIdAndUpdate(contactId, body, { new: true }).where({ ownerId: userId });
 };
 
 module.exports = {
